Renamed company import export folder
